+++ project/ps-service/src/main/java/com/codeshare/permission/product/dao/IProductFileDao.java
@@ -3,13 +3,13 @@ package com.codeshare.permission.product.dao;
 import com.codeshare.permission.product.po.ProductFile;
 
 public interface IProductFileDao {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(ProductFile record);
 
     int insertSelective(ProductFile record);
 
-    ProductFile selectByPrimaryKey(Integer id);
+    ProductFile selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(ProductFile record);
