+++ project/ps-service/src/main/java/com/codeshare/permission/product/dao/IProductPriceDao.java
@@ -3,13 +3,13 @@ package com.codeshare.permission.product.dao;
 import com.codeshare.permission.product.po.ProductPrice;
 
 public interface IProductPriceDao {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(ProductPrice record);
 
     int insertSelective(ProductPrice record);
 
-    ProductPrice selectByPrimaryKey(Integer id);
+    ProductPrice selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(ProductPrice record);
